# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHIC Engineering Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List all issues with filters
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: limit
          schema: { type: number }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: repositoryId
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: severity
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: classification
          schema: { type: string }
        - in: query
          name: assignee
          schema: { type: string }
        - in: query
          name: label
          schema: { type: string }
        - in: query
          name: aiStatus
          schema: { type: string }
      responses:
        "200":
          description: Issues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueListResponse"
  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get issue details with AI analysis
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"
  /issues/{id}/bulk-update:
    patch:
      operationId: bulkUpdateIssues
      tags: [issues]
      summary: Bulk update issues
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  # ─── Commits ─────────────────────────────────────────────────────────────────
...
  /duplicates:
    get:
      operationId: listDuplicateClusters
      tags: [duplicates]
      summary: List duplicate issue clusters
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: repositoryId
          schema: { type: string }
        - in: query
          name: minSimilarity
          schema: { type: number }
      responses:
        "200":
          description: Duplicate clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateClusterListResponse"

  # ─── Regressions ──────────────────────────────────────────────────────────────
  /regressions:
    get:
      operationId: listRegressions
      tags: [regressions]
      summary: List regressions
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: repositoryId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: confidence
          schema: { type: string }
      responses:
        "200":
          description: Regressions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegressionListResponse"

  # ─── Automation ───────────────────────────────────────────────────────────────
  /automation/suggestions:
    get:
      operationId: listAutomationSuggestions
      tags: [automation]
      summary: List AI automation suggestions
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: repositoryId
          schema: { type: string }
      responses:
        "200":
          description: Automation suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationSuggestionListResponse"
  /automation/suggestions/{id}/approve:
    post:
      operationId: approveAutomationSuggestion
      tags: [automation]
      summary: Approve an automation suggestion
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /automation/suggestions/{id}/reject:
    post:
      operationId: rejectAutomationSuggestion
      tags: [automation]
      summary: Reject an automation suggestion
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ─── Engineering Intelligence ─────────────────────────────────────────────────
  /intelligence/summary:
    get:
      operationId: getIntelligenceSummary
      tags: [intelligence]
      summary: Get engineering intelligence summary
      parameters:
        - in: query
          name: repositoryId
          schema: { type: string }
      responses:
        "200":
          description: Intelligence summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntelligenceSummary"
  /intelligence/root-causes:
    get:
      operationId: listRootCauses
      tags: [intelligence]
      summary: Root cause explorer
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: repositoryId
          schema: { type: string }
      responses:
        "200":
          description: Root causes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RootCauseEntry"
  /intelligence/historical-patterns:
    get:
      operationId: listHistoricalPatterns
      tags: [intelligence]
      summary: Historical patterns
      parameters:
        - in: query
          name: repositoryId
          schema: { type: string }
        - in: query
          name: period
          schema: { type: string }
      responses:
        "200":
          description: Historical patterns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoricalPattern"
  /intelligence/risk-analysis:
    get:
      operationId: getRiskAnalysis
      tags: [intelligence]
      summary: Risk analysis across repositories
      parameters:
        - in: query
          name: repositoryId
          schema: { type: string }
      responses:
        "200":
          description: Risk analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskAnalysis"

  # ─── Analytics ────────────────────────────────────────────────────────────────
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview across all repositories
      parameters:
        - in: query
          name: repositoryId
          schema: { type: string }
        - in: query
          name: period
          schema: { type: string }
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"

  # ─── Search ───────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: unreadOnly
          schema: { type: boolean }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ─── Integrations ─────────────────────────────────────────────────────────────
...
  /api-keys:
    get:
      operationId: listApiKeys
      tags: [apikeys]
      summary: List API keys
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"

    post:
      operationId: createApiKey
      tags: [apikeys]
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
  /api-keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [apikeys]
      summary: Delete an API key
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  # ─── Organization ─────────────────────────────────────────────────────────────
...
  /organization/members:
    get:
      operationId: listMembers
      tags: [organization]
      summary: List organization members
      parameters:
        - in: query
          name: page
          schema: { type: number }
        - in: query
          name: role
          schema: { type: string }
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListResponse"
  /organization/members/{id}:
    patch:
      operationId: updateMember
      tags: [organization]
      summary: Update member role
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"

  # ─── Settings ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
      required: [success, message]

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
    TimeSeriesPoint:
      type: object
      properties:
        date: { type: string }
        value: { type: number }
      required: [date, value]
...
    Issue:
      type: object
      properties:
        id: { type: string }
        number: { type: number }
        repositoryId: { type: string }
        repositoryName: { type: string }
        title: { type: string }
        status: { type: string }
        priority: { type: string }
        severity: { type: string }
        classification: { type: string }
        actionability: { type: string }
        confidence: { type: number }
        labels:
          type: array
          items: { type: string }
        assignee: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        aiStatus: { type: string }
        hasEngineeringReport: { type: boolean }
        url: { type: string }
      required:
        - id
        - number
        - repositoryId
        - repositoryName
        - title
        - status
        - priority
        - severity
        - classification
        - actionability
        - confidence
        - labels
        - createdAt
        - updatedAt
        - aiStatus
        - hasEngineeringReport
        - url
    IssueListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Issue"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]
    IssueUpdate:
      type: object
      properties:
        priority: { type: string }
        severity: { type: string }
        assignee: { type: string }
        status: { type: string }
        labels:
          type: array
          items: { type: string }
    IssueDetail:
      type: object
      properties:
        issue:
          $ref: "#/components/schemas/Issue"
        executiveSummary: { type: string }
        rootCauseAnalysis: { type: ["string", "null"] }
        regressionSignals: { type: ["string", "null"] }
        engineeringImpact: { type: ["string", "null"] }
        investigationPlan: { type: ["string", "null"] }
        suggestedTests:
          type: array
          items: { type: string }
        suggestedLabels:
          type: array
          items: { type: string }
        suggestedAssignee: { type: ["string", "null"] }
        missingInformation: { type: ["string", "null"] }
        relevantFiles:
          type: array
          items: { type: string }
        relevantFunctions:
          type: array
          items: { type: string }
        relevantComponents:
          type: array
          items: { type: string }
        relatedCommits:
          type: array
          items:
            $ref: "#/components/schemas/CommitRef"
        relatedPullRequests:
          type: array
          items:
            $ref: "#/components/schemas/PullRequestRef"
        relatedIssues:
          type: array
          items:
            $ref: "#/components/schemas/IssueRef"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        automationSuggestions:
          type: array
          items:
            $ref: "#/components/schemas/AutomationSuggestion"
      required:
        - issue
        - executiveSummary
        - relevantFiles
        - relevantFunctions
        - relevantComponents
        - relatedCommits
        - relatedPullRequests
        - relatedIssues
        - timeline
        - automationSuggestions
        - suggestedTests
        - suggestedLabels
    CommitRef:
      type: object
      properties:
        sha: { type: string }
        message: { type: string }
        author: { type: string }
        date: { type: string }
      required: [sha, message, author, date]
    PullRequestRef:
      type: object
      properties:
        number: { type: number }
        title: { type: string }
        status: { type: string }
        url: { type: string }
      required: [number, title, status, url]
    IssueRef:
      type: object
      properties:
        id: { type: string }
        number: { type: number }
        title: { type: string }
        status: { type: string }
      required: [id, number, title, status]
    TimelineEvent:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        actor: { type: string }
        description: { type: string }
        timestamp: { type: string }
      required: [id, type, actor, description, timestamp]

    # ─── Commit ───────────────────────────────────────────────────────────────
...
    DuplicateCluster:
      type: object
      properties:
        id: { type: string }
        primaryIssue:
          $ref: "#/components/schemas/IssueRef"
        possibleDuplicates:
          type: array
          items:
            $ref: "#/components/schemas/DuplicateCandidate"
        sharedComponents:
          type: array
          items: { type: string }
        sharedFiles:
          type: array
          items: { type: string }
        sharedCommits:
          type: array
          items: { type: string }
        recommendedAction: { type: string }
        createdAt: { type: string }
      required:
        - id
        - primaryIssue
        - possibleDuplicates
        - sharedComponents
        - sharedFiles
        - sharedCommits
        - recommendedAction
        - createdAt
...
    DuplicateClusterListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DuplicateCluster"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Regression ───────────────────────────────────────────────────────────
    Regression:
      type: object
      properties:
        id: { type: string }
        issue:
          $ref: "#/components/schemas/IssueRef"
        confidence: { type: number }
        relatedCommit: { type: ["string", "null"] }
        relatedCommitSha: { type: ["string", "null"] }
        relatedRelease: { type: ["string", "null"] }
        affectedFiles:
          type: array
          items: { type: string }
        regressionWindowStart: { type: string }
        regressionWindowEnd: { type: string }
        investigationStatus: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        detectedAt: { type: string }
      required:
        - id
        - issue
        - confidence
        - affectedFiles
        - regressionWindowStart
        - regressionWindowEnd
        - investigationStatus
        - repositoryId
        - repositoryName
        - detectedAt
    RegressionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Regression"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Automation ───────────────────────────────────────────────────────────
    AutomationSuggestion:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        issueId: { type: ["string", "null"] }
        issueTitle: { type: ["string", "null"] }
        repositoryId: { type: string }
        repositoryName: { type: string }
        content: { type: string }
        status: { type: string }
        confidence: { type: number }
        createdAt: { type: string }
        reviewedAt: { type: ["string", "null"] }
        reviewedBy: { type: ["string", "null"] }
      required:
        - id
        - type
        - repositoryId
        - repositoryName
        - content
        - status
        - confidence
        - createdAt
    AutomationSuggestionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AutomationSuggestion"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]

    # ─── Intelligence ─────────────────────────────────────────────────────────
    IntelligenceSummary:
      type: object
      properties:
        totalRootCauses: { type: number }
        criticalPatterns: { type: number }
        activeRegressions: { type: number }
        knowledgeItems: { type: number }
        riskLevel: { type: string }
        topInsight: { type: ["string", "null"] }
        engineeringHealthTrend: { type: string }
        lastAnalysisAt: { type: ["string", "null"] }
      required:
        - totalRootCauses
        - criticalPatterns
        - activeRegressions
        - knowledgeItems
        - riskLevel
        - engineeringHealthTrend
    RootCauseEntry:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        description: { type: string }
        affectedIssues: { type: number }
        repositoryId: { type: string }
        repositoryName: { type: string }
        confidence: { type: number }
        firstSeenAt: { type: string }
        lastSeenAt: { type: string }
      required:
        - id
        - category
        - description
        - affectedIssues
        - repositoryId
        - repositoryName
        - confidence
        - firstSeenAt
        - lastSeenAt
    HistoricalPattern:
      type: object
      properties:
        id: { type: string }
        patternType: { type: string }
        description: { type: string }
        frequency: { type: string }
        repositoryId: { type: string }
        repositoryName: { type: string }
        impact: { type: string }
        occurrences: { type: number }
        firstSeenAt: { type: string }
        lastSeenAt: { type: string }
      required:
        - id
        - patternType
        - description
        - frequency
        - repositoryId
        - repositoryName
        - impact
        - occurrences
        - firstSeenAt
        - lastSeenAt
    RiskAnalysis:
      type: object
      properties:
        overallRisk: { type: string }
        overallRiskScore: { type: number }
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/RepositoryRisk"
        topRisks:
          type: array
          items:
            $ref: "#/components/schemas/RiskItem"
        trendDirection: { type: string }
        lastCalculatedAt: { type: string }
      required:
        - overallRisk
        - overallRiskScore
        - repositories
        - topRisks
        - trendDirection
        - lastCalculatedAt
    RepositoryRisk:
      type: object
      properties:
        repositoryId: { type: string }
        repositoryName: { type: string }
        riskScore: { type: number }
        riskLevel: { type: string }
        topRiskFactor: { type: string }
      required: [repositoryId, repositoryName, riskScore, riskLevel, topRiskFactor]
    RiskItem:
      type: object
      properties:
        category: { type: string }
        description: { type: string }
        severity: { type: string }
        repositoryName: { type: string }
      required: [category, description, severity, repositoryName]

    # ─── Analytics Overview ───────────────────────────────────────────────────
    AnalyticsOverview:
      type: object
      properties:
        period: { type: string }
        totalIssues: { type: number }
        resolvedIssues: { type: number }
        openIssues: { type: number }
        regressions: { type: number }
        duplicatesFound: { type: number }
        avgResolutionDays: { type: number }
        issuesByPriority:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        issuesBySeverity:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        issuesByClassification:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        issuesByRepository:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/TimeSeriesPoint"
      required:
        - period
        - totalIssues
        - resolvedIssues
        - openIssues
        - regressions
        - duplicatesFound
        - avgResolutionDays
        - issuesByPriority
        - issuesBySeverity
        - issuesByClassification
        - issuesByRepository
        - trend
...
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total: { type: number }
        unread: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, unread, page, limit]

    # ─── Integration ──────────────────────────────────────────────────────────
...
    ApiKey:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        prefix: { type: string }
        scopes:
          type: array
          items: { type: string }
        createdAt: { type: string }
        lastUsedAt: { type: ["string", "null"] }
        expiresAt: { type: ["string", "null"] }
        rateLimit: { type: number }
        usageCount: { type: number }
      required: [id, name, prefix, scopes, createdAt, rateLimit, usageCount]
    ApiKeyInput:
      type: object
      properties:
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
        expiresAt: { type: string }
        rateLimit: { type: number }
      required: [name, scopes]

    # ─── Organization ─────────────────────────────────────────────────────────
...
    Member:
      type: object
      properties:
        id: { type: string }
        login: { type: string }
        name: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string }
        joinedAt: { type: string }
        lastActiveAt: { type: ["string", "null"] }
      required: [id, login, role, joinedAt]
    MemberListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
      required: [items, total, page, limit]
    MemberUpdate:
      type: object
      properties:
        role: { type: string }

    # ─── Settings ─────────────────────────────────────────────────────────────
...